Continue porting UploadService and NetRequest classes. Added default values for json ’get’ functions

+++ Smile Identity Demo/smile_id_sdk/core/upload/SIDResponse.swift
@@ -11,6 +11,6 @@ import Foundation
 class SIDResponse {
     var partnerParams   : PartnerParams?
     var success         : Bool?
-    var confidenceValue : Int?
+    var confidenceValue : Float?
       
 }
